feat: Support for Capacitor 5
