+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "@nativescript/capacitor",
-  "version": "4.0.0",
+  "version": "5.0.0-beta.0",
   "description": "NativeScript for Capacitor",
   "main": "dist/plugin.cjs.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/esm/index.d.ts",
+  "module": "dist/esm/index.mjs",
+  "types": "dist/esm/index.d.mts",
   "unpkg": "dist/plugin.js",
   "files": [
     "android/src/main/",
@@ -19,9 +19,9 @@
     "src/nativescript/"
   ],
   "bin": {
-    "build-nativescript": "./bin/build-nativescript.js",
-    "dev-nativescript": "./bin/dev-nativescript.js",
-    "uninstall-nativescript": "./bin/uninstall-nativescript.js"
+    "build-nativescript": "./bin/build-nativescript.mjs",
+    "dev-nativescript": "./bin/dev-nativescript.mjs",
+    "uninstall-nativescript": "./bin/uninstall-nativescript.mjs"
   },
   "author": "NativeScript TSC",
   "license": "Apache-2.0",
@@ -48,51 +48,51 @@
     "prettier": "prettier \"**/*.{css,html,ts,js,java}\"",
     "swiftlint": "node-swiftlint",
     "docgen": "docgen --api NativeScriptCapPlugin --output-readme README.md --output-json dist/docs.json",
-    "build": "npm run clean && npm run docgen && tsc && rollup -c rollup.config.js && npm run build:bridge",
-    "build:bridge": "cd src_bridge && ts-patch install && tsc && cp package.json ../bridge/package.json && cp webpack.config.js ../bridge/webpack.config.js && cp webpack4.config.js ../bridge/webpack4.config.js && cp legacy-ns-transform-native-classes.js ../bridge/legacy-ns-transform-native-classes.js && cp native-custom.d.ts ../bridge/native-custom.d.ts",
+    "build": "npm run clean && npm run docgen && tsc && rollup -c rollup.config.js --bundleConfigAsCjs && npm run build:bridge",
+    "build:bridge": "cd src_bridge && ts-patch install && tsc && cp package.json ../bridge/package.json && cp webpack.config.js ../bridge/webpack.config.js && cp webpack4.config.js ../bridge/webpack4.config.js && cp legacy-ns-transform-native-classes.js ../bridge/legacy-ns-transform-native-classes.js && cp native-custom.d.ts ../bridge/native-custom.d.ts && cp ../src/nativeapi.d.ts ../dist/esm/nativeapi.d.ts",
     "clean": "rimraf ./dist",
     "clean.all": "npx rimraf dist node_modules package-lock.json yarn.lock && npm run setup",
     "setup": "npm i --legacy-peer-deps --ignore-scripts && ts-patch install",
     "watch": "tsc --watch",
     "prepublishOnly": "npm run build",
-    "postinstall": "node dist/esm/postinstall --action install"
+    "postinstall": "node dist/esm/postinstall.mjs --action install"
   },
   "dependencies": {
     "@nativescript/webpack": "~5.0.0",
     "chokidar": "^3.5.2",
-    "fs-extra": "^9.1.0",
-    "nativescript-dev-xcode": "~0.3.0",
+    "fs-extra": "^11.1.0",
+    "nativescript-dev-xcode": "~0.5.0",
     "npm-run-all": "~4.1.5",
     "prompts": "^2.4.1",
-    "strip-json-comments": "^3.1.1",
-    "ts-loader": "^9.2.6",
+    "strip-json-comments": "^5.0.0",
+    "ts-loader": "^9.4.2",
     "ts-node": "^10.9.0",
-    "ts-patch": "^2.0.0",
+    "ts-patch": "^2.1.0",
     "unzipper": "^0.10.11",
     "xmldom": "^0.6.0",
-    "yargs": "^17.4.0",
-    "webpack-cli": "^4.9.2"
+    "yargs": "17.6.0"
   },
   "devDependencies": {
-    "@capacitor/android": "^4.3.0",
-    "@capacitor/core": "^4.3.0",
-    "@capacitor/docgen": "^0.2.0",
-    "@capacitor/ios": "^4.3.0",
+    "@capacitor/android": "^5.0.1",
+    "@capacitor/core": "^5.0.1",
+    "@capacitor/docgen": "0.2.1",
+    "@capacitor/ios": "^5.0.1",
     "@ionic/eslint-config": "^0.3.0",
-    "@ionic/prettier-config": "^2.0.0",
+    "@ionic/prettier-config": "^3.0.0",
     "@ionic/swiftlint-config": "^1.1.2",
-    "@nativescript/core": "~8.3.0",
-    "@nativescript/types": "~8.3.0",
-    "@types/fs-extra": "^9.0.6",
+    "@nativescript/core": "~8.5.0",
+    "@nativescript/types": "~8.5.0",
+    "@types/fs-extra": "^11.0.1",
     "@types/node": "^18.0.0",
     "@types/xmldom": "^0.1.30",
-    "eslint": "8.15.0",
-    "prettier": "~2.6.2",
-    "prettier-plugin-java": "~1.6.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.56.0",
-    "swiftlint": "^1.0.1",
-    "typescript": "~4.7.3"
+    "@types/yargs": "^17.0.24",
+    "eslint": "^8.40.0",
+    "prettier": "~2.8.8",
+    "prettier-plugin-java": "~2.1.0",
+    "rimraf": "^5.0.0",
+    "rollup": "^3.21.5",
+    "swiftlint": "^1.0.2",
+    "typescript": "~4.9.0"
   },
   "prettier": "@ionic/prettier-config",
   "swiftlint": "@ionic/swiftlint-config",
