+++ src/nativescript/package.json
@@ -1,7 +1,7 @@
 {
   "name": "nativescript",
   "dependencies": {
-    "@nativescript/core": "~8.3.0",
-    "@nativescript/types": "~8.3.0"
+    "@nativescript/core": "~8.5.0",
+    "@nativescript/types": "~8.5.0"
   }
 }
