+++ src/nativescript/references.d.ts
@@ -1,6 +1,6 @@
 /// <reference path="./node_modules/@nativescript/core/global-types.d.ts" />
 /// <reference path="./node_modules/@nativescript/types-ios/index.d.ts" />
-/// <reference path="./node_modules/@nativescript/types-android/lib/android-29.d.ts" />
+/// <reference path="./node_modules/@nativescript/types-android/lib/android-33.d.ts" />
 
 import type { nativeCustom } from '../native-custom';
