+++ src_bridge/references.d.ts
@@ -1,6 +1,6 @@
 /// <reference path="../node_modules/@nativescript/core/global-types.d.ts" />
 /// <reference path="../node_modules/@nativescript/types-ios/index.d.ts" />
-/// <reference path="../node_modules/@nativescript/types-android/lib/android-30.d.ts" />
+/// <reference path="../node_modules/@nativescript/types-android/lib/android-33.d.ts" />
 
 declare module NodeJS  {
   interface Global {
