+++ tsconfig.json
@@ -3,18 +3,16 @@
     "allowUnreachableCode": false,
     "declaration": true,
     "esModuleInterop": true,
-    "lib": [
-      "dom",
-      "es2017"
-    ],
+    "lib": ["ESNext", "dom"],
     "module": "esnext",
+    "target": "es2020",
     "moduleResolution": "node",
     "experimentalDecorators": true,
     "noFallthroughCasesInSwitch": true,
+    "removeComments": false,
     "outDir": "dist/esm",
     "pretty": true,
     "sourceMap": true,
-    "target": "es2017",
     "plugins": [
       {
         "transform": "@nativescript/webpack/dist/transformers/NativeClass/index",
@@ -23,7 +21,7 @@
     ]
   },
   "files": [
-    "src/index.ts",
-    "src/postinstall.ts"
+    "src/index.mts",
+    "src/postinstall.mts"
   ]
 }
